requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coverage==4.5.1
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-JWT-Extended==3.13.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
psycopg2-binary==2.7.5
py==1.7.0
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.9.1
pytest-cov==2.6.0
python-dotenv==0.9.1
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
